Export booleans for even better errors
